+++ src/index.js
@@ -60,7 +60,7 @@ const install = app => {
 }
 
 const tyhUi = {
-  version: "1.4.6",
+  version: "1.4.7",
   install
 }